requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
certifi==2024.7.4
click==8.1.7
Django==5.0.7
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.1.0
h11==0.14.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
PySocks==1.7.1
pytest==8.2.2
pytest-django==4.8.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
